student admission & teacher list

+++ routes/web.php
@@ -15,12 +15,15 @@ use App\Http\Controllers\SessionController;
 use App\Http\Controllers\SettingController;
 use App\Http\Controllers\ShopController;
 use App\Http\Controllers\Auth\AuthenticatedSessionController;
+use App\Http\Controllers\ChildrenController;
+use App\Http\Controllers\ProgrammeController;
 use App\Http\Controllers\StudentController;
+use App\Http\Controllers\TeacherController;
 
 /* Authorized Only */
-Route::middleware(['auth', 'check_role'])->group(function(){
+Route::middleware(['auth'])->group(function(){
     /* Default Landing Page */
-    Route::get('/', [DashboardController::class, 'index']);
+    Route::get('/', [DashboardController::class, 'index'])->middleware('permission:dashboard_view_access');
     // Route::get('/', function () {
     //     return Inertia::render('Welcome', [
     //         'canLogin' => Route::has('login'),
@@ -30,20 +33,27 @@ Route::middleware(['auth', 'check_role'])->group(function(){
     //     ]);
     // });
     /* User Impersonation */
-    Route::get('/impersonate/{user}', [AuthenticatedSessionController::class, 'impersonate'])->name('impersonate');
-    Route::get('/leave-impersonate', [AuthenticatedSessionController::class, 'leaveImpersonate'])->name('leave-impersonate');
+    Route::get('/impersonate/{user}', [AuthenticatedSessionController::class, 'impersonate'])->name('impersonate')->middleware('permission:impersonate_access|is_impersonated');
+    Route::get('/leave-impersonate', [AuthenticatedSessionController::class, 'leaveImpersonate'])->name('leave-impersonate')->middleware('permission:impersonate_access|is_impersonated');;
 
     /* Dashboard */
-        /* User */
-        Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
+    /* User */
+    Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard')->middleware('permission:dashboard_view_access');
 
     /* Profile */
     Route::get('/profile', [ProfileController::class, 'create'])->name('profile');
     Route::post('/profile/store', [ProfileController::class, 'store'])->name('profile.store');
     Route::post('/profile/security/store', [ProfileController::class, 'storeSecurity'])->name('profile.security.store');
 
-    /* Users */
+    /* Children */
+    Route::get('/children', [ChildrenController::class, 'create'])->name('children');
+    Route::post('/children/store', [ChildrenController::class, 'store'])->name('children.store');
+    Route::get('/children/edit', [ChildrenController::class, 'edit'])->name('children.edit');
+    Route::post('/children/update', [ChildrenController::class, 'update'])->name('children.update');
+
+    /* Control Panel */
     Route::prefix('control-panel')->group(function () {
+        /* Users */
         Route::get('/users', [UserController::class, 'index'])->name('users');
         Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
         Route::post('/users/store', [UserController::class, 'store'])->name('users.store');
@@ -53,12 +63,12 @@ Route::middleware(['auth', 'check_role'])->group(function(){
         Route::post('/users/completed-tour', [UserController::class, 'completedTour'])->name('users.completed_tour');
 
         /* Roles */
-        Route::get('/roles', [RoleController::class, 'index'])->name('roles');
-        Route::get('/roles/create', [RoleController::class, 'create'])->name('roles.create');
-        Route::post('/roles/store', [RoleController::class, 'store'])->name('roles.store');
-        Route::get('/roles/edit', [RoleController::class, 'edit'])->name('roles.edit');
-        Route::post('/roles/update', [RoleController::class, 'update'])->name('roles.update');
-        Route::delete('/roles/destroy/{role}', [RoleController::class, 'destroy'])->name('roles.destroy');
+        Route::get('/roles', [RoleController::class, 'index'])->name('roles')->middleware('permission:cp_roles_view_access');
+        Route::get('/roles/create', [RoleController::class, 'create'])->name('roles.create')->middleware('permission:cp_roles_create_access');
+        Route::post('/roles/store', [RoleController::class, 'store'])->name('roles.store')->middleware('permission:cp_roles_create_access');
+        Route::get('/roles/edit', [RoleController::class, 'edit'])->name('roles.edit')->middleware('permission:cp_roles_edit_access');
+        Route::post('/roles/update', [RoleController::class, 'update'])->name('roles.update')->middleware('permission:cp_roles_edit_access');
+        Route::delete('/roles/destroy/{role}', [RoleController::class, 'destroy'])->name('roles.destroy')->middleware('permission:cp_roles_delete_access');
         Route::get('/roles/assign-permissions', [RoleController::class, 'assignPermissions'])->name('roles.assign_pemissions');
         Route::post('/roles/assign-permissions/store', [RoleController::class, 'assignPermissionsStore'])->name('roles.assign_permissions.store');
 
@@ -88,73 +98,14 @@ Route::middleware(['auth', 'check_role'])->group(function(){
         Route::delete('/permissions/destroy/{id}', [PermissionController::class, 'destroy'])->name('permissions.destroy');
 
         /* Countries */
-        Route::get('/control-panel/general/countries', [SettingController::class, 'countryList'])->name('settings.countries');
-        Route::get('/control-panel/general/countries/create', [SettingController::class, 'addCountry'])->name('settings.countries.create');
-        Route::post('/control-panel/general/countries/store', [SettingController::class, 'storeCountry'])->name('settings.countries.store');
-        Route::get('/control-panel/general/countries/edit', [SettingController::class, 'editCountry'])->name('settings.countries.edit');
-        Route::post('/control-panel/general/countries/update', [SettingController::class, 'updateCountry'])->name('settings.countries.update');
-        Route::delete('/control-panel/general/countries/destroy/{id}', [SettingController::class, 'destroyCountry'])->name('settings.countries.destroy');
-    });
+        Route::get('/general/countries', [SettingController::class, 'countryList'])->name('settings.countries');
+        Route::get('/general/countries/create', [SettingController::class, 'addCountry'])->name('settings.countries.create');
+        Route::post('/general/countries/store', [SettingController::class, 'storeCountry'])->name('settings.countries.store');
+        Route::get('/general/countries/edit', [SettingController::class, 'editCountry'])->name('settings.countries.edit');
+        Route::post('/general/countries/update', [SettingController::class, 'updateCountry'])->name('settings.countries.update');
+        Route::delete('/general/countries/destroy/{id}', [SettingController::class, 'destroyCountry'])->name('settings.countries.destroy');
 
-    /* Centres */
-    Route::prefix('centre-management')->group(function () {
-        Route::get('/centres', [CentreController::class, 'index'])->name('centres');
-        Route::get('/centres/create', [CentreController::class, 'create'])->name('centres.create');
-        Route::post('/centres/store', [CentreController::class, 'store'])->name('centres.store');
-        Route::get('/centres/edit', [CentreController::class, 'edit'])->name('centres.edit');
-        Route::post('/centres/update', [CentreController::class, 'update'])->name('centres.update');  
-        Route::delete('/centres/destroy/{id}', [CentreController::class, 'destroy'])->name('centres.destroy');
-        Route::delete('/centres/image/destroy/{id}', [CentreController::class, 'destroyImage'])->name('centres.destroy_image');
-        Route::get('/centres/images', [CentreController::class, 'getCentreImages'])->name('centres.get_images');
-
-        /* Classes */
-        Route::get('/classes', [ClassController::class, 'index'])->name('classes');
-        Route::get('/classes/create', [ClassController::class, 'create'])->name('classes.create');
-        Route::post('/classes/store', [ClassController::class, 'store'])->name('classes.store');
-        Route::get('/classes/edit', [ClassController::class, 'edit'])->name('classes.edit');
-        Route::post('/classes/update', [ClassController::class, 'update'])->name('classes.update');  
-        Route::delete('/classes/destroy/{id}', [ClassController::class, 'destroy'])->name('classes.destroy');
-        Route::get('/classes/get-class-types', [ClassController::class, 'getClassTypes'])->name('classes.get_class_types');
-        Route::get('/classes/get-class-levels', [ClassController::class, 'getClassLevels'])->name('classes.get_class_levels');
-        Route::get('/classes/find', [ClassController::class, 'findClasses'])->name('classes.find');
-
-        /* Students */
-        Route::get('/students', [StudentController::class, 'index'])->name('students');
-        Route::get('/students/create', [StudentController::class, 'create'])->name('students.create');
-        Route::post('/students/store', [StudentController::class, 'store'])->name('students.store');
-        Route::get('/students/edit', [StudentController::class, 'edit'])->name('students.edit');
-        Route::post('/students/update', [StudentController::class, 'update'])->name('students.update');  
-        Route::delete('/students/destroy/{id}', [StudentController::class, 'destroy'])->name('students.destroy');
-        Route::get('/students/find', [StudentController::class, 'findStudents'])->name('students.find');
-
-        /* Settings */    
-        Route::get('/settings', [SettingController::class, 'programmeList'])->name('settings');
-            /* Class Types */
-            Route::get('/settings/class_types', [SettingController::class, 'classTypeList'])->name('settings.class_types');
-            Route::get('/settings/class_types/create', [SettingController::class, 'addClassType'])->name('settings.class_types.create');
-            Route::post('/settings/class_types/store', [SettingController::class, 'storeClassType'])->name('settings.class_types.store');
-            Route::get('/settings/class_types/edit', [SettingController::class, 'editClassType'])->name('settings.class_types.edit');
-            Route::post('/settings/class_types/update', [SettingController::class, 'updateClassType'])->name('settings.class_types.update');
-            Route::delete('/settings/class_types/destroy/{id}', [SettingController::class, 'destroyClassType'])->name('settings.class_types.destroy');
-
-            /* Programmes*/
-            Route::get('/settings/programmes', [SettingController::class, 'programmeList'])->name('settings.programmes');
-            Route::get('/settings/programmes/create', [SettingController::class, 'addProgramme'])->name('settings.programmes.create');
-            Route::post('/settings/programmes/store', [SettingController::class, 'storeProgramme'])->name('settings.programmes.store');
-            Route::get('/settings/programmes/edit', [SettingController::class, 'editProgramme'])->name('settings.programmes.edit');
-            Route::post('/settings/programmes/update', [SettingController::class, 'updateProgramme'])->name('settings.programmes.update');  
-            Route::delete('/settings/programmes/destroy/{id}', [SettingController::class, 'destroyProgramme'])->name('settings.programmes.destroy');
-            Route::get('/settings/programmes/get-fee', [SettingController::class, 'getFee'])->name('settings.get_fee');
-    });
-
-
-
-    /* Run Diagnostic Test */
-    Route::get('/diagnostic-test', [DiagnosticTestController::class, 'dtStartList'])->name('diagnostic_test');
-    Route::get('/diagnostic-test/start', [DiagnosticTestController::class, 'dtStart'])->name('diagnostic_test.run');
-
-    /* Diagnostic Test Setup*/
-    Route::prefix('control-panel')->group(function () {
+        // DT Setup
         /* Diagnostic Test List */
         Route::get('/general/diagnostic-test', [SettingController::class, 'dtList'])->name('settings.diagnostic_test');
         Route::get('/general/diagnostic-test/create', [SettingController::class, 'dtCreate'])->name('settings.diagnostic_test.create');
@@ -186,6 +137,67 @@ Route::middleware(['auth', 'check_role'])->group(function(){
         Route::post('/general/diagnostic-test/categories/update', [SettingController::class, 'dtCategoriesUpdate'])->name('settings.diagnostic_test.categories.update');
         Route::delete('/general/diagnostic-test/categories/destroy/{id}', [SettingController::class, 'dtCategoriesDestroy'])->name('settings.diagnostic_test.categories.destroy');
     });
+
+    /* Centre Management */
+    Route::prefix('centre-management')->group(function () {
+        /* Centres */
+        Route::get('/centres', [CentreController::class, 'index'])->name('centres')->middleware('permission:centre_view_access');
+        Route::get('/centres/create', [CentreController::class, 'create'])->name('centres.create')->middleware('permission:centre_create_access');
+        Route::post('/centres/store', [CentreController::class, 'store'])->name('centres.store')->middleware('permission:centre_create_access');
+        Route::get('/centres/edit', [CentreController::class, 'edit'])->name('centres.edit')->middleware('permission:centre_edit_access');
+        Route::post('/centres/update', [CentreController::class, 'update'])->name('centres.update')->middleware('permission:centre_edit_access');  
+        Route::delete('/centres/destroy/{id}', [CentreController::class, 'destroy'])->name('centres.destroy')->middleware('permission:centre_delete_access');
+        Route::delete('/centres/image/destroy/{id}', [CentreController::class, 'destroyImage'])->name('centres.destroy_image')->middleware('permission:centre_delete_access');
+        Route::get('/centres/images', [CentreController::class, 'getCentreImages'])->name('centres.get_images')->middleware('permission:centre_view_access');
+        Route::get('/centres/find-user', [CentreController::class, 'findUser'])->name('centres.find_user')->middleware('permission:centre_view_access');
+
+        /* Classes */
+        Route::get('/classes', [ClassController::class, 'index'])->name('classes')->middleware('permission:class_view_access');
+        Route::get('/classes/create', [ClassController::class, 'create'])->name('classes.create')->middleware('permission:class_create_access');
+        Route::post('/classes/store', [ClassController::class, 'store'])->name('classes.store');
+        Route::get('/classes/edit', [ClassController::class, 'edit'])->name('classes.edit');
+        Route::post('/classes/update', [ClassController::class, 'update'])->name('classes.update');  
+        Route::delete('/classes/destroy/{id}', [ClassController::class, 'destroy'])->name('classes.destroy');
+        Route::get('/classes/get-class-types', [ClassController::class, 'getClassTypes'])->name('classes.get_class_types');
+        Route::get('/classes/get-class-levels', [ClassController::class, 'getClassLevels'])->name('classes.get_class_levels');
+        Route::get('/classes/find', [ClassController::class, 'findClasses'])->name('classes.find');
+
+        /* Students */
+        Route::get('/students', [StudentController::class, 'index'])->name('students');
+        Route::get('/students/create', [StudentController::class, 'create'])->name('students.create');
+        Route::post('/students/store', [StudentController::class, 'store'])->name('students.store');
+        Route::get('/students/edit', [StudentController::class, 'edit'])->name('students.edit');
+        Route::post('/students/update', [StudentController::class, 'update'])->name('students.update');
+        Route::delete('/students/destroy', [StudentController::class, 'destroy'])->name('students.destroy');
+        Route::get('/students/find', [StudentController::class, 'findStudents'])->name('students.find');
+        Route::post('/students/add-student-class', [StudentController::class, 'addStudentClass'])->name('students.add_student_class');
+
+        /* Teachers */
+        Route::get('/teachers', [TeacherController::class, 'index'])->name('teachers');
+
+        /* Settings */    
+        Route::get('/settings', [SettingController::class, 'programmeList'])->name('settings');
+            /* Class Types */
+            Route::get('/settings/class_types', [SettingController::class, 'classTypeList'])->name('settings.class_types');
+            Route::get('/settings/class_types/create', [SettingController::class, 'addClassType'])->name('settings.class_types.create');
+            Route::post('/settings/class_types/store', [SettingController::class, 'storeClassType'])->name('settings.class_types.store');
+            Route::get('/settings/class_types/edit', [SettingController::class, 'editClassType'])->name('settings.class_types.edit');
+            Route::post('/settings/class_types/update', [SettingController::class, 'updateClassType'])->name('settings.class_types.update');
+            Route::delete('/settings/class_types/destroy/{id}', [SettingController::class, 'destroyClassType'])->name('settings.class_types.destroy');
+
+            /* Programmes*/
+            Route::get('/settings/programmes', [SettingController::class, 'programmeList'])->name('settings.programmes');
+            Route::get('/settings/programmes/create', [SettingController::class, 'addProgramme'])->name('settings.programmes.create');
+            Route::post('/settings/programmes/store', [SettingController::class, 'storeProgramme'])->name('settings.programmes.store');
+            Route::get('/settings/programmes/edit', [SettingController::class, 'editProgramme'])->name('settings.programmes.edit');
+            Route::post('/settings/programmes/update', [SettingController::class, 'updateProgramme'])->name('settings.programmes.update');  
+            Route::delete('/settings/programmes/destroy/{id}', [SettingController::class, 'destroyProgramme'])->name('settings.programmes.destroy');
+            Route::get('/settings/programmes/get-fee', [SettingController::class, 'getFee'])->name('settings.get_fee');
+    });
+
+    /* Run Diagnostic Test */
+    Route::get('/diagnostic-test', [DiagnosticTestController::class, 'dtStartList'])->name('diagnostic_test');
+    Route::get('/diagnostic-test/start', [DiagnosticTestController::class, 'dtStart'])->name('diagnostic_test.run');
     
     /* Shop */
     Route::get('/shop', [ShopController::class, 'index'])->name('shop');
@@ -193,5 +205,8 @@ Route::middleware(['auth', 'check_role'])->group(function(){
     Route::get('/shop/checkout', [ShopController::class, 'shopCheckout'])->name('shop.checkout');
     Route::post('/shop/payment', [ShopController::class, 'shopPayment'])->name('shop.payment');
     Route::get('/shop/payment/status', [ShopController::class, 'shopPaymentStatus'])->name('shop.payment.status');
+
+    /* Web notifications */
+    Route::get('/send_web_notifications', [ProgrammeController::class, 'send_web_notifications'])->name('send_web_notifications');
 });
 require __DIR__.'/auth.php';
